Annotate splice junctions with resulting transcript sequence

Annotate splice junctions with resulting transcript sequence

add_context_seq(df, transcripts, size = 400, bsg = NULL, keep_ranges = FALSE)

Arguments

df

A data.frame with splice junctions in rows and at least the columns:

junc_id junction id consisting of genomic coordinates

tx_id the ID of the affected transcript (see add_tx)

transcripts

as a named GRangesList of transcripts

size

the size of the output sequence around the junction position (might be shorter if transcripts is shorter)

bsg

BSgenome object such as BSgenome.Hsapiens.UCSC.hg19

keep_ranges

Should GRanges of transcripts and modified transcript be kept? If TRUE, the list columns tx_lst and tx_mod_lst are added to the output.

Value

A data.frame with the same rows as the input df but with the following additional column(s):

tx_mod_id an identifier made from tx_id and junc_id

junc_pos_tx the junction position in the modified transcript sequence

cts_seq the context sequence

cts_junc_pos the junction position in the context sequence

cts_size the size of the context sequence

cts_id a unique id for the context sequence as hash value using the XXH128 hash algorithm

If the keep_ranges is TRUE, the following additional columns are added to the output data.frame:

tx_lst a list of GRanges with the original transcript as provided in tx_id column and transcripts object..

tx_mod_lst a list of GRanges with the modified transcript (see modify_tx)

Examples

requireNamespace("BSgenome.Hsapiens.UCSC.hg19", quietly = TRUE)
bsg <- BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19

add_context_seq(toy_junc_df, toy_transcripts, size = 20, bsg = bsg)
#> # A tibble: 14 × 8
#>    junc_id  tx_id  tx_mod_id   junc_pos_tx cts_seq cts_junc_pos cts_size cts_id 
#>    <chr>    <chr>  <chr>             <int> <chr>          <dbl>    <int> <chr>  
#>  1 chr2_15… ENST0… ENST000004…       16412 GATGAA…           10       20 a8a77f…
#>  2 chr2_15… ENST0… ENST000004…       16517 GATCAG…           10       20 5fcb4b…
#>  3 chr2_15… ENST0… ENST000004…       17290 GTGGAG…           10       20 5eba9f…
#>  4 chr2_15… ENST0… ENST000004…       16412 GATGAA…           10       20 87853e…
#>  5 chr2_15… ENST0… ENST000004…       16517 GATCAG…           10       20 36d94e…
#>  6 chr2_17… ENST0… ENST000003…       83789 TCTCCA…           10       20 eae0ac…
#>  7 chr2_17… ENST0… ENST000003…       84158 AATTTT…           10       20 8c710f…
#>  8 chr2_17… ENST0… ENST000003…       83789 TCTCCA…           10       20 2cc1c2…
#>  9 chr2_17… ENST0… ENST000003…       59307 ACCAGG…           10       20 1a8ca2…
#> 10 chr2_17… ENST0… ENST000003…       59288 GACATC…           10       20 77864e…
#> 11 chr2_17… ENST0… ENST000003…       58982 GACCCT…           10       20 64eab4…
#> 12 chr2_17… ENST0… ENST000003…        4828 CCAAAA…           10       20 7f12a2…
#> 13 chr2_17… ENST0… ENST000003…        4868 ACTGTG…           10       20 d4a800…
#> 14 chr2_17… ENST0… ENST000003…        4703 TTTCAT…           10       20 350f25…