Annotate splice junctions with resulting CDS and peptide sequence

Annotate splice junctions with resulting CDS and peptide sequence

add_peptide(df, size = 30, bsg = NULL)

Arguments

df
A data.frame with splice junctions in rows and at least the colums:

junc_id junction id consisting of genomic coordinates

tx_id the ID of the affected transcript (see add_tx)

cds_lst a list of GRanges with the CDS (see add_tx)

size
the total size of the output sequence (might be shorter if peptide is shorter)

bsg
BSgenome object such as BSgenome.Hsapiens.UCSC.hg19

Value

A data.frame as the input with the additional column(s):

cds_alt_lst a list of GRanges with the modified CDS ranges.

cds_id_alt an identifier made from tx_id and junc_id

junc_pos_cds the junction position in the modified CDS sequence

protein the full protein sequence of the translated modified CDS.

protein_junc_pos The position of the junctino in the protein sequence

junc_in_orf Indicator whehter the junction is located in an open reading frame.

peptide_context_seq_raw the peptide sequence around the junction includig stop codons.

peptide_context the peptide sequence around the junction truncated after stop codons.

peptide_context_junc_pos The junction position relative to the peptide_context sequence

Examples

requireNamespace("BSgenome.Hsapiens.UCSC.hg19", quietly = TRUE)
bsg <- BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19

junc_df <- toy_junc_df %>%
  dplyr::mutate(
    cds_lst = as.list(toy_cds[tx_id])
  )

add_peptide(junc_df, size = 30, bsg = bsg)
#> # A tibble: 14 × 12
#>    junc_id   tx_id  cds_lst  cds_alt_lst cds_id_alt   junc_pos_cds protein      
#>    <chr>     <chr>  <named > <named lis> <chr>               <int> <chr>        
#>  1 chr2_152… ENST0… <GRange… <GRanges>   ENST0000040…        16209 MADDEDYEEVVE…
#>  2 chr2_152… ENST0… <GRange… <GRanges>   ENST0000040…        16314 MADDEDYEEVVE…
#>  3 chr2_152… ENST0… <GRange… <GRanges>   ENST0000040…        17087 MADDEDYEEVVE…
#>  4 chr2_152… ENST0… <GRange… <GRanges>   ENST0000040…        16209 MADDEDYEEVVE…
#>  5 chr2_152… ENST0… <GRange… <GRanges>   ENST0000040…        16314 MADDEDYEEVVE…
#>  6 chr2_179… ENST0… <GRange… <GRanges>   ENST0000034…        83566 MTTQAPTFTQPL…
#>  7 chr2_179… ENST0… <GRange… <GRanges>   ENST0000034…        83935 MTTQAPTFTQPL…
#>  8 chr2_179… ENST0… <GRange… <GRanges>   ENST0000034…        83566 MTTQAPTFTQPL…
#>  9 chr2_179… ENST0… <GRange… <GRanges>   ENST0000034…        59084 MTTQAPTFTQPL…
#> 10 chr2_179… ENST0… <GRange… <GRanges>   ENST0000034…        59065 MTTQAPTFTQPL…
#> 11 chr2_179… ENST0… <GRange… <GRanges>   ENST0000034…        58759 MTTQAPTFTQPL…
#> 12 chr2_179… ENST0… <GRange… <GRanges>   ENST0000034…         4605 MTTQAPTFTQPL…
#> 13 chr2_179… ENST0… <GRange… <GRanges>   ENST0000034…         4645 MTTQAPTFTQPL…
#> 14 chr2_179… ENST0… <GRange… <GRanges>   ENST0000034…         4480 MTTQAPTFTQPL…
#> # … with 5 more variables: protein_junc_pos <dbl>, junc_in_orf <lgl>,
#> #   peptide_context_seq_raw <chr>, peptide_context <chr>,
#> #   peptide_context_junc_pos <dbl>