Calculate distance of splice site to next canonical splice site in the transcript.

Calculate distance of splice site to next canonical splice site in the transcript.

calc_distance_regular(
  pos,
  tx_lst,
  pos_exon_idx,
  pos_exon_boundary,
  site = "left"
)

Arguments

pos

A GRanges representation of the splice junction.

tx_lst

A list of GRanges with the transcript.

pos_exon_idx

Exon index in transcript if splice site overlaps exon.

pos_exon_boundary

Exon index in transcript if junction matches exon boundary.

site

A string describing the site of the junction to be annotated.

Value

An integer of the genomic distance in bp.