Select a subset of transcripts per junction that are more likely to be affected by a junction.

All possible affected transcripts are required and can be annotated with add_tx()

choose_tx(df)

Arguments

df

A data.frame with splice junctions in rows and at least the columns:

junc_id junction id consisting of genomic coordinates

tx_id transcript id of possibly affected transcripts

tx_lst a list of GRanges with the transcript

Value

A data.frame as with relevant transcript and junction combinations. If tx_id is NA in the input data.frame such rows are removed from the output data.frame.

This function selects transcripts that are more likely to be affected to reduce the amount of junction and transcript combinations. The function excludes transcripts for which both junction positions are located in an intron. Junctions in a given transcript must either represent an exon skipping, intron retention, exitron, or alternative splice site event or have both junction positions in an exon. Other junction-transcript combinations are also excluded. This function may loose relevant or keep irrelevant junction-transcripts in particular in regions with multiple isoforms with distinct splicing pattern.

See also

Examples

junc_df <- tibble::tibble(
  junc_id = c("chr2:152389996-152392205:-", "chr2:152389996-152390729:-",
              "chr2:152389955-152389956:-")
)

junc_df <- add_tx(junc_df, toy_transcripts)
choose_tx(junc_df)
#> # A tibble: 19 × 6
#>    junc_id            tx_id tx_lst    putative_event_type distance_to_next_can…¹
#>    <chr>              <chr> <named l> <chr>                                <dbl>
#>  1 chr2:152389996-15… ENST… <GRanges> ASS                                      0
#>  2 chr2:152389996-15… ENST… <GRanges> ASS                                      0
#>  3 chr2:152389996-15… ENST… <GRanges> ASS                                      0
#>  4 chr2:152389996-15… ENST… <GRanges> ASS                                      0
#>  5 chr2:152389996-15… ENST… <GRanges> ASS                                      0
#>  6 chr2:152389996-15… ENST… <GRanges> ASS                                      0
#>  7 chr2:152389996-15… ENST… <GRanges> ASS                                      0
#>  8 chr2:152389996-15… ENST… <GRanges> ASS                                      0
#>  9 chr2:152389996-15… ENST… <GRanges> ASS                                      0
#> 10 chr2:152389996-15… ENST… <GRanges> ASS                                      0
#> 11 chr2:152389996-15… ENST… <GRanges> ASS                                      0
#> 12 chr2:152389996-15… ENST… <GRanges> ASS                                      0
#> 13 chr2:152389996-15… ENST… <GRanges> ASS                                      0
#> 14 chr2:152389996-15… ENST… <GRanges> ASS                                      0
#> 15 chr2:152389996-15… ENST… <GRanges> ASS                                      0
#> 16 chr2:152389955-15… ENST… <GRanges> IR                                       0
#> 17 chr2:152389955-15… ENST… <GRanges> IR                                       0
#> 18 chr2:152389955-15… ENST… <GRanges> IR                                       0
#> 19 chr2:152389955-15… ENST… <GRanges> IR                                       0
#> # ℹ abbreviated name: ¹​distance_to_next_canonical_donor
#> # ℹ 1 more variable: distance_to_next_canonical_acceptor <dbl>