Formats pangolin output and filter for predicted effects

Reformat the data for each annotated effect per row. As pangolin does not provide information on donor or acceptor, use both for gain and loss annotations. Effects with score = 0 are filtered out.

format_pangolin(variants)

Arguments

variants

tibble with parsed pangolin mutations from parse_pangolin

Value

A tibble with splicing effects per row

See also

Examples

pangolin_file <- system.file("extdata", "spliceai_output.pangolin.vcf", package = "splice2neo")
df <- parse_pangolin(pangolin_file)
format_pangolin(df)
#> # A tibble: 18 × 6
#>    mut_id                 effect score chr   pos_rel       pos
#>    <chr>                  <chr>  <dbl> <chr>   <int>     <int>
#>  1 chr1_25000_A_C,G,T     DG      0.01 chr1        8     25008
#>  2 chr1_25000_A_C,G,T     AG      0.01 chr1        8     25008
#>  3 chr2_152389953_T_A,C,G DG      0.01 chr2        2 152389955
#>  4 chr2_152389953_T_A,C,G AG      0.01 chr2        2 152389955
#>  5 chr2_179415988_C_CA    DG      0.08 chr2      -26 179415962
#>  6 chr2_179415988_C_CA    AG      0.08 chr2      -26 179415962
#>  7 chr2_179446218_ATACT_A DG      0.08 chr2       -7 179446211
#>  8 chr2_179446218_ATACT_A AG      0.08 chr2       -7 179446211
#>  9 chr2_179642185_G_A     DG      0.54 chr2        2 179642187
#> 10 chr2_179642185_G_A     AG      0.54 chr2        2 179642187
#> 11 chr19_38958362_C_T     DG      0.85 chr19      -2  38958360
#> 12 chr19_38958362_C_T     AG      0.85 chr19      -2  38958360
#> 13 chr21_47406854_CCA_C   DG      0.06 chr21     -38  47406816
#> 14 chr21_47406854_CCA_C   AG      0.06 chr21     -38  47406816
#> 15 chr21_47406856_A_AT    DG      0.08 chr21     -40  47406816
#> 16 chr21_47406856_A_AT    AG      0.08 chr21     -40  47406816
#> 17 chrX_129274636_A_C,G,T DG      0.01 chrX      -41 129274595
#> 18 chrX_129274636_A_C,G,T AG      0.01 chrX      -41 129274595