Formats spliceAI output and filter for predicted effects

Reformat the data for each annotated effect per row, filters effects to have a probability not NA and score > 0, and removes gene symbol from data to make non-redundant output.

format_spliceai(spliceai_variants)

Arguments

spliceai_variants

tibble with parsed spliceAI mutations from parse_spliceai

Value

A tibble with splicing effects per row

See also

Examples

spliceai_file <- system.file("extdata", "spliceai_output.vcf", package = "splice2neo")
df <- parse_spliceai(spliceai_file)
format_spliceai(df)
#> # A tibble: 21 × 12
#>    CHROM POS    ID    REF   ALT   QUAL  FILTER   Key ALLELE change  prob pos_rel
#>    <chr> <chr>  <chr> <chr> <chr> <chr> <chr>  <int> <chr>  <fct>  <dbl>   <int>
#>  1 chr2  15238… NA    T     A,C,G NA    NA         2 A      AG      0.01      43
#>  2 chr2  15238… NA    T     A,C,G NA    NA         2 A      DL      0.74       3
#>  3 chr2  15238… NA    T     A,C,G NA    NA         2 C      AG      0.04      43
#>  4 chr2  15238… NA    T     A,C,G NA    NA         2 C      DL      0.71       3
#>  5 chr2  15238… NA    T     A,C,G NA    NA         2 G      AG      0.03      43
#>  6 chr2  15238… NA    T     A,C,G NA    NA         2 G      DL      0.75       3
#>  7 chr2  17941… NA    C     CA    NA    NA         3 CA     AG      0.07      -7
#>  8 chr2  17941… NA    C     CA    NA    NA         3 CA     AL      1         -1
#>  9 chr2  17944… NA    ATACT A     NA    NA         4 A      DG      0.02     -11
#> 10 chr2  17944… NA    ATACT A     NA    NA         4 A      DL      0.91       8
#> # … with 11 more rows