Combines tibbles with junctions from any number of RNA-seq tools into a combined dataset of expressed splice junctions

Combines tibbles with junctions from any number of RNA-seq tools into a combined dataset of expressed splice junctions

generate_combined_dataset(rna_junc_data_list)

Arguments

rna_junc_data_list

A named list with junction tibbles in standardized format.

Value

A combined table with unique junctions. The columns identified_by_{name} contains information which tools identified the given junction

Examples

path <-  system.file("extdata", "", package = "splice2neo")
spladder_juncs <- spladder_transform(path)
#> Importing Spladder files ...
#> Importing Spladder files completed
path <-  system.file("extdata", "test_regtools_Aligned.out.sorted.bam.junc", package = "splice2neo")
regtools_juncs <- regtools_transform(path)
dat.combined <- generate_combined_dataset(list("spladder" = spladder_juncs,
                                               "regtools" = regtools_juncs))