Tests if start or and end of junc is on exon.

Tests if start or and end of junc is on exon.

is_start_on_exon(tx_lst, jx_lst, intron_retention)

Arguments

tx_lst

transcripts a GRangesList with transcripts defined as GRanges of exons.

jx_lst

splice junctions as GRanges objects in list format

intron_retention

logical vector indicating if junctions belong to an intron retention event

Value

a logical vector indicating if start or and end of junc is on exon.