modifies transcripts by applying splice-junctions

modifies transcripts by applying splice-junctions

modify_tx(tx, jx)

Arguments

tx
transcripts a named GRangesList with transcripts defined as GRanges of exons.

jx
splice junctions as GRanges objects with the same length as tx.

Value

a GRangesList with altered transcripts.

The algorithm works in the following steps.

Get transcripts for which the whole transcript range contains pos1 and pos2 as GRangesList of GRanges of exons

For each tx in transcripts do

if pos1 amd pos2 are on the same exon, duplicate exon and consider the first for pos1 and the second for pos2.

if pos1 overlaps an exon e: shorten e to end at pos1 else: get next exon e left of pos: extend e to end in pos1

if pos2 overlaps an exon e: shorten e to start at pos2 else: get next exon e right of pos2: extend e to start in pos2

remove all exons completeley contained in the interval [pos1, pos2].

Examples

tx <- GenomicRanges::GRangesList(
  list(GenomicRanges::GRanges(c(
     "1:2-3:+",
     "1:5-6:+",
     "1:10-15:+"
  )))
)

jx <- GenomicRanges::GRanges(c("1:3-10:+"))

modify_tx(tx, jx)
#> GRangesList object of length 1:
#> [[1]]
#> GRanges object with 2 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]        1       2-3      +
#>   [2]        1     10-15      +
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#>