Imports SplAdder output from a given path and transforms it into standardized junction format. Results from one patients should be stored per folder.

Imports SplAdder output from a given path and transforms it into standardized junction format. Results from one patients should be stored per folder.

spladder_transform(path)

Arguments

path

The path to a folder with SplAdder output

Value

A tibble in standardized junction format, combining all alternative splicing classes that are covered by SplAdder

Examples

path <-  system.file("extdata", "", package = "splice2neo")
spladder_juncs <- spladder_transform(path)
#> Importing Spladder files ...
#> Importing Spladder files completed
spladder_juncs
#> # A tibble: 26 × 8
#>    junction_start junction_end strand chromosome Gene              class junc_id
#>    <chr>          <chr>        <chr>  <chr>      <chr>             <chr> <chr>  
#>  1 896180         896673       +      chr1       ENSG00000187961.… A3SS  chr1:8…
#>  2 896180         897009       +      chr1       ENSG00000187961.… A3SS  chr1:8…
#>  3 889462         891303       -      chr1       ENSG00000188976.… A3SS  chr1:8…
#>  4 889903         891303       -      chr1       ENSG00000188976.… A3SS  chr1:8…
#>  5 763229         764383       +      chr1       ENSG00000228794.… A5SS  chr1:7…
#>  6 763233         764383       +      chr1       ENSG00000228794.… A5SS  chr1:7…
#>  7 934993         935072       -      chr1       ENSG00000188290.… A5SS  chr1:9…
#>  8 934993         935246       -      chr1       ENSG00000188290.… A5SS  chr1:9…
#>  9 764484         776580       +      chr1       ENSG00000228794.… cass… chr1:7…
#> 10 776753         783034       +      chr1       ENSG00000228794.… cass… chr1:7…
#> # ℹ 16 more rows
#> # ℹ 1 more variable: number_supporting_reads <dbl>