Transforms Spladder output into standardized junction format

Transforms Spladder output into standardized junction format

spladder_transform_format(l)

Arguments

l
A list with tibbles that contain the spladder output - each for one type of alternative splicing. These types can be "A5SS", "A3SS", "cassette_exon", "intron_retention", "mutex_exons".

Value

A tibble in standardized junction format, combining all alternative splicing classes that are were determined with Spladder

Examples

spladder_output
#> $A3SS
#> # A tibble: 1 × 15
#>   contig strand event_id     gene_name           exon_const_start exon_const_end
#>   <chr>  <chr>  <chr>        <chr>               <chr>            <chr>         
#> 1 chr1   -      alt_3prime.1 ENSG00000171163.16… 21878165         21878359      
#> # … with 9 more variables: exon_alt1_start <chr>, exon_alt1_end <chr>,
#> #   exon_alt2_start <chr>, exon_alt2_end <chr>,
#> #   sample_Aligned.out_sorted:exon_diff_cov <dbl>,
#> #   sample_Aligned.out_sorted:exon_const_cov <dbl>,
#> #   sample_Aligned.out_sorted:intron1_conf <dbl>,
#> #   sample_Aligned.out_sorted:intron2_conf <dbl>,
#> #   sample_Aligned.out_sorted:psi <lgl>
#> 
#> $A5SS
#> # A tibble: 1 × 15
#>   contig strand event_id     gene_name           exon_const_start exon_const_end
#>   <chr>  <chr>  <chr>        <chr>               <chr>            <chr>         
#> 1 chr1   -      alt_5prime.1 ENSG00000171163.16… 59228292         59228349      
#> # … with 9 more variables: exon_alt1_start <chr>, exon_alt1_end <chr>,
#> #   exon_alt2_start <chr>, exon_alt2_end <chr>,
#> #   sample_Aligned.out_sorted:exon_diff_cov <dbl>,
#> #   sample_Aligned.out_sorted:exon_const_cov <dbl>,
#> #   sample_Aligned.out_sorted:intron1_conf <dbl>,
#> #   sample_Aligned.out_sorted:intron2_conf <dbl>,
#> #   sample_Aligned.out_sorted:psi <lgl>
#> 
#> $cassette_exon
#> # A tibble: 1 × 15
#>   contig strand event_id     gene_name    exon_pre_start exon_pre_end exon_start
#>   <chr>  <chr>  <chr>        <chr>        <chr>          <chr>        <chr>     
#> 1 chr1   +      alt_5prime.1 ENSG0000017… 59228292       59228349     59228549  
#> # … with 8 more variables: exon_end <chr>, exon_aft_start <chr>,
#> #   exon_aft_end <chr>, sample_Aligned.out_sorted:exon_diff_cov <dbl>,
#> #   sample_Aligned.out_sorted:exon_const_cov <dbl>,
#> #   sample_Aligned.out_sorted:intron1_conf <dbl>,
#> #   sample_Aligned.out_sorted:intron2_conf <dbl>,
#> #   sample_Aligned.out_sorted:psi <lgl>
#> 
#> $intron_retention
#> # A tibble: 1 × 15
#>   contig strand event_id gene_name exon1_start exon1_end intron_start intron_end
#>   <chr>  <chr>  <chr>    <chr>     <chr>       <chr>     <chr>        <chr>     
#> 1 chr1   -      alt_5pr… ENSG0000… 244974      249445    249446       249512    
#> # … with 7 more variables: exon2_start <chr>, exon2_end <chr>,
#> #   sample_Aligned.out_sorted:exon_diff_cov <dbl>,
#> #   sample_Aligned.out_sorted:exon_const_cov <dbl>,
#> #   sample_Aligned.out_sorted:intron1_conf <dbl>,
#> #   sample_Aligned.out_sorted:intron2_conf <dbl>,
#> #   sample_Aligned.out_sorted:psi <lgl>
#> 
#> $mutex_exons
#> # A tibble: 1 × 17
#>   contig strand event_id     gene_name   exon_pre_start exon_pre_end exon1_start
#>   <chr>  <chr>  <chr>        <chr>       <chr>          <chr>        <chr>      
#> 1 chr1   -      alt_5prime.1 ENSG000001… 244974         249445       249446     
#> # … with 10 more variables: exon1_end <chr>, exon2_start <chr>,
#> #   exon2_end <chr>, exon_aft_start <chr>, exon_aft_end <chr>,
#> #   sample_Aligned.out_sorted:exon_diff_cov <dbl>,
#> #   sample_Aligned.out_sorted:exon_const_cov <dbl>,
#> #   sample_Aligned.out_sorted:intron1_conf <dbl>,
#> #   sample_Aligned.out_sorted:intron2_conf <dbl>,
#> #   sample_Aligned.out_sorted:psi <lgl>
#> 
transformed_spladder <- spladder_transform_format(
 spladder_output)
transformed_spladder
#> # A tibble: 12 × 8
#>    junction_start junction_end strand chromosome Gene  class AS_event_ID junc_id
#>             <dbl>        <dbl> <chr>  <chr>      <chr> <chr> <chr>       <chr>  
#>  1       21877890     21878165 -      chr1       ENSG… A3SS  alt_3prime… chr1_2…
#>  2       21877997     21878165 -      chr1       ENSG… A3SS  alt_3prime… chr1_2…
#>  3       59228349     59222127 -      chr1       ENSG… A5SS  alt_5prime… chr1_5…
#>  4       59228349     59222127 -      chr1       ENSG… A5SS  alt_5prime… chr1_5…
#>  5       59228349     59228549 +      chr1       ENSG… cass… alt_5prime… chr1_5…
#>  6       59228749     59229049 +      chr1       ENSG… cass… alt_5prime… chr1_5…
#>  7       59228349     59229049 +      chr1       ENSG… cass… alt_5prime… chr1_5…
#>  8         249445       249446 -      chr1       ENSG… intr… alt_5prime… chr1_2…
#>  9         249512       249513 -      chr1       ENSG… intr… alt_5prime… chr1_2…
#> 10         249445       249446 -      chr1       ENSG… mute… alt_5prime… chr1_2…
#> 11         249512       249513 -      chr1       ENSG… mute… alt_5prime… chr1_2…
#> 12         249631       249513 -      chr1       ENSG… mute… alt_5prime… chr1_2…