Creates a table with context sequences for re-quantification analysis with Easyquant.

Creates a table with context sequences for re-quantification analysis with Easyquant.

transform_for_requant(cts_id, cts_seq, junc_position)

Arguments

cts_id
An id for the cts sequence to be re-quantified. If the context sequence was determined with the function junct_to_cts the associated hash id cts_id can be used.

cts_seq
The context sequence.

junc_position
The position of the junction within the context sequence. If the context sequence was determined with junct_to_cts the associated hash id cts_junc_pos can be used.

Value

A tibble that can be used for requantification with easyquant. This tibble has the columns

name: Name of the input sequence.

sequence: The context sequence that will be requantified.

position: The position of the junction within the context sequence.

Examples