Annotate splice junctions with resulting transcript sequence

Annotate splice junctions with resulting transcript sequence

add_context_seq(df, size = 400, bsg = NULL)

Arguments

df
A data.frame with splice junctions in rows and at least the colums:

junc_id junction id consisting of genomic coordinates

tx_id the ID of the affected transcript (see add_tx)

tx_lst a list of GRanges with the transcript (see add_tx)

size
the size of the output sequence around the junction position (might be shorter if transcripts is shorter)

bsg
BSgenome object such as BSgenome.Hsapiens.UCSC.hg19

Value

A data.frame as the input with the additional column(s):

tx_alt_lst a list of GRanges with the modified transcript (see modify_tx)

tx_id_alt an identifier made from tx_id and junc_id

junc_pos_tx the junction position in the modified transcript sequence

cts_seq the context sequence

cts_junc_pos the junction position in the context sequence

cts_size the size of the context sequence

cts_id a unique id for the context sequence as hash value using the XXH128 hash algorithm

Examples

requireNamespace("BSgenome.Hsapiens.UCSC.hg19", quietly = TRUE)
bsg <- BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19

junc_df <- toy_junc_df %>%
  dplyr::mutate(
    tx_lst = as.list(toy_transcripts[tx_id])
  )

add_context_seq(junc_df, size = 30, bsg = bsg)
#> # A tibble: 14 × 10
#>    junc_id  tx_id  tx_lst tx_alt_lst tx_id_alt  junc_pos_tx cts_seq cts_junc_pos
#>    <chr>    <chr>  <name> <named li> <chr>            <int> <chr>          <dbl>
#>  1 chr2_15… ENST0… <GRan… <GRanges>  ENST00000…       16412 CAATTG…           14
#>  2 chr2_15… ENST0… <GRan… <GRanges>  ENST00000…       16517 GACCAG…           14
#>  3 chr2_15… ENST0… <GRan… <GRanges>  ENST00000…       17290 CATATG…           14
#>  4 chr2_15… ENST0… <GRan… <GRanges>  ENST00000…       16412 CAATTG…           14
#>  5 chr2_15… ENST0… <GRan… <GRanges>  ENST00000…       16517 GACCAG…           14
#>  6 chr2_17… ENST0… <GRan… <GRanges>  ENST00000…       83789 CTGTTT…           14
#>  7 chr2_17… ENST0… <GRan… <GRanges>  ENST00000…       84158 CTTATA…           14
#>  8 chr2_17… ENST0… <GRan… <GRanges>  ENST00000…       83789 CTGTTT…           14
#>  9 chr2_17… ENST0… <GRan… <GRanges>  ENST00000…       59307 GTATTA…           14
#> 10 chr2_17… ENST0… <GRan… <GRanges>  ENST00000…       59288 CCAAGG…           14
#> 11 chr2_17… ENST0… <GRan… <GRanges>  ENST00000…       58982 CTCGGG…           14
#> 12 chr2_17… ENST0… <GRan… <GRanges>  ENST00000…        4828 GTTGCC…           14
#> 13 chr2_17… ENST0… <GRan… <GRanges>  ENST00000…        4868 TTTTAA…           14
#> 14 chr2_17… ENST0… <GRan… <GRanges>  ENST00000…        4703 TCTGGT…           14
#> # … with 2 more variables: cts_size <int>, cts_id <chr>